shared/requirements.txt
sqlalchemy
pgvector
psycopg2-binary
sentence-transformers
PyJWT
docker
